requirements.txt
pymupdf 
PyPDF2 
easyocr 
sentence_transformers 
faiss-cpu 
langchain-core 
langchain-upstage 
langchain-community 
numpy 
rank-bm25 
scikit-learn 
bert_score 
rouge_score 
nltk
ragas 
datasets 
openai
python-dotenv
cohere